agent: prevent prompt provider FD leak
The parent process must close the FD passed to a child if it no longer
needs it. This commit does so by using the already existing scope object
to close the FD as soon as the scope is left.

This makes Mir notices when the trust prompt is disappearing, which
fixes ubports/ubuntu-touch#1346 by giving the focus back to the app as
soon as possible.