mir/agent: fix indefinite wait if fail to start prompt session
For some reason, Mir still provide an FD for a prompt session even if it
fails to start. When this FD is passed to the prompt provider, it will
just hang, causing the whole trust store user chain to just wait. To fix
this, make sure the agent checks for the error before continue.

Related: https://github.com/ubports/ubuntu-touch/issues/1668