mir/agent: query UAL for the application's primary PID
Mir expects a PID of a process that have a Mir session. Meanwhile, the
trust-store users supply us the PID of the requesting process, which
might not be the same process. For example, QtWebEngine uses a helper
process to record the audio. Thus, we have to find the PID of the app's
main PID, which hopefully has a Mir session.

Fixes: https://github.com/ubports/ubuntu-touch/issues/1668