mir/agent: re-try prompt session with parent
Mir expects a PID of a process that have a Mir session. Meanwhile, the
trust-store users supply us the PID of the requesting process, which
might not be the same process. For example, QtWebEngine uses a helper
process to record the audio. In this case, Mir won't be able to find a
session and fails.

What we do here is that, when we encounter such error from Mir, get the
process's parent pid, and then try again.

Thanks @mariogrip for the idea.

Fixes: https://github.com/ubports/morph-browser/issues/399
Fixes: https://github.com/ubports/ubuntu-touch/issues/1668